refactor: 💡 Switch to yarn 3 and implement monorepo structure with workspaces
This commit updates the project to use yarn 3 as the package manager
instead of NPM. Additionally, I have implemented a monorepo structure
that leverages yarn workspaces to manage both frontend and backend
projects.  With this setup, can share common dependencies between
projects, manage versioning easily, and run scripts across all projects
in the workspace.  To implement this change, I have updated the
project's package.json to include a "workspaces" field that specifies
the location of the sub-projects, and also added a new root-level
yarn.lock file to manage dependencies across the workspace.  Overall,
this change should improve the maintainability and scalability of
project, as well as simplify the development workflow for me.